requirements.txt
kivy
wand
pillow
tensorflow
tensorflow_hub
waifu2x-vulkan